% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{data}
\name{abc}
\alias{abc}
\title{ABC function}
\format{
An object of class \code{list} of length 4.
}
\usage{
abc
}
\arguments{
\item{observedData}{Observed data (e.g., as an Armadillo matrix or vector)}

\item{numParticles}{Number of particles/samples to generate}

\item{numIters}{Number of iterations}

\item{epsilon}{Tolerance threshold}
}
\value{
posteriorSamples Generated posterior parameter samples (e.g., as an Armadillo matrix)
}
\description{
ABC function
}
\keyword{datasets}
