% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeSummaryStatistics}
\alias{computeSummaryStatistics}
\title{Function to compute summary statistics for simulated data from the zombie outbreak model}
\usage{
computeSummaryStatistics(simulatedData)
}
\arguments{
\item{simulatedData}{A 3D array of simulated data with dimensions (numParticles, numTimePoints, numPopStats)}
}
\value{
A matrix of summary statistics with dimensions (numParticles, numStats). Each row represents the summary statistics for a particular particle. The number of columns (numStats) depends on the selected summary statistics
}
\description{
Function to compute summary statistics for simulated data from the zombie outbreak model
}
