% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimatePosterior}
\alias{estimatePosterior}
\title{Estimate the posterior distribution based on accepted parameter samples and weights}
\arguments{
\item{acceptedSamples}{Matrix of accepted parameter samples (each row represents a set of accepted parameter values)}

\item{weights}{Vector of weights assigned to accepted parameter samples}
}
\value{
posteriorSamples Matrix of posterior parameter samples
}
\description{
Estimate the posterior distribution based on accepted parameter samples and weights
}
