% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate.SZR.data}
\alias{generate.SZR.data}
\title{Gives simulated data for a deterministic SZR model}
\usage{
generate.SZR.data(
  N = NULL,
  initial.inf = NULL,
  total.T = NULL,
  stochastic = TRUE
)
}
\arguments{
\item{N}{The total population}

\item{initial.inf}{The number of infected individuals at time point 1}

\item{total.T}{The number of time points to simulate}
}
\value{
A list with elements
- `results` A data frame with `total.T` rows and 4 columns showing the population counts for each time point
- `N` The total population, as input
- `beta` The true bite rate
- `kappa` The true kill rate
- `rho` The true resurrection rate
}
\description{
Gives simulated data for a deterministic SZR model
}
