% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateParameterSamples}
\alias{generateParameterSamples}
\title{Generate parameter samples from the prior}
\arguments{
\item{num_Params}{Number of parameters}

\item{numParticles}{Number of parameter samples to generate}

\item{priorMin}{Lower bound of the prior distribution}

\item{priorMax}{Upper bound of the prior distribution}
}
\value{
parameterSamples Generated parameter samples as an Armadillo matrix
}
\description{
Generate parameter samples from the prior
}
