% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{performInference}
\alias{performInference}
\title{Perform posterior inference and compute posterior summaries}
\usage{
performInference(posteriorSamples)
}
\arguments{
\item{posteriorSamples}{Matrix of posterior parameter samples (each row represents a set of posterior parameter values)}
}
\value{
posteriorMean Vector of posterior means for each parameter

posteriorMedian Vector of posterior medians for each parameter

posteriorQuantiles Matrix of posterior quantiles for each parameter (columns represent quantiles)

posteriorCI Matrix of posterior confidence intervals for each parameter (columns represent lower and upper bounds)
}
\description{
Perform posterior inference and compute posterior summaries
}
